WDX_256_P3: add virtual referencing prototype
Summary:
Prototype code for virtual referencing. If the snapshot file has an impedance
file in the same directory, this widget will now import it, decide which
channels are "good" (min and max settings at the top of the script), average
over those channels, and then subtract this reference signal from all channels
as they're displayed. This is toggled by pressing the V key, and indicated by
the addition of '(virtual)' to the window title.

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>

Test Plan: Justin tested at BU

Reviewers: #willow, cblam

Reviewed By: #willow, cblam

Subscribers: #willow

Differential Revision: https://reviews.leaflabs.com/D428