WillowDataset: add default args to importSlice()
Summary:
Add default time sample arguments to importSlice so that, when it is
called with no arguments, it processes a slice across the whole length
of time of the selected dataset (much like importData()).

Test Plan:
Does not affect previous functionality of importSlice().
(This is part of a stack of commits which will allow probe-mapped plots
to be run again.)

Reviewers: #willow, cblam

Reviewed By: #willow, cblam

Subscribers: #willow

Differential Revision: https://reviews.leaflabs.com/D417