WillowDataset: add chans option arg to importSlice
Summary:
Add an optional parameter, 'chans', to WillowDataset's importSlice()
method, which allows the selection of a subset of channels upon slice
importation. When 'chans' is not given, all channels are imported,
maintaining the same functionality in importSlice() as before.

Depends on D417.

Test Plan:
Does not affect previous functionality of importSlice().
(This is part of a stack of commits which will allow probe-mapped plots
to be run again.)

Reviewers: #willow, cblam

Reviewed By: #willow, cblam

Subscribers: #willow

Differential Revision: https://reviews.leaflabs.com/D418