WillowDataset: add some attrs for filtered slices
Summary:
Add attributes 'self.slice_filtered_min' and 'self.slice_filtered_max',
created upon filtering a slice of data. These attributes are analagous
to the previously-existing 'self.slice_min' and 'self.slice_max', which
are created upon importing a slice of data.

Depends on D418.

Test Plan:
Does not affect previous functionality of filterSlice().
(This is part of a stack of commits which will allow probe-mapped plots
to be run again.)

Reviewers: #willow, cblam

Reviewed By: #willow, cblam

Subscribers: #willow

Differential Revision: https://reviews.leaflabs.com/D419