WillowDataset: resurrect spike-detection method
Summary:
Re-implement the function detectSpikes() from previous versions of
WillowDataset, altering its functionality to work on slices rather than
whole datasets, and changing the default argument 'thresh' to be a real
default argument instead of needing to be specified with a string.

Depends on D425.

Test Plan:
after this commit you'll be able to test by seeing that the
SpikeScopeWindow works again

Reviewers: #willow, cblam

Reviewed By: #willow, cblam

Subscribers: #willow

Differential Revision: https://reviews.leaflabs.com/D426